+++ crates/cli/tests/reference/enums.d.ts
@@ -37,6 +37,12 @@ export enum Color {
    */
   Red = 2,
 }
+export enum ImplicitDiscriminant {
+  A = 0,
+  B = 1,
+  C = 42,
+  D = 43,
+}
 /**
  * The name of a color.
  */